feat(network): implement Phase 4 -- WiFi, BLE, HTTP server, SSE, and embedded WebUI
Add complete network subsystem for ESP32: WifiManager (AP/STA/captive
portal/UDP DNS), BleManager (NUS GATT, 3-level zombie defense), HTTP
server (12 REST routes + SSE streaming), embedded WebUI dashboard with
captive portal, and main-loop transport SM.

- WifiManager: AP mode (192.168.4.1/24), STA with NVS credential
  persistence, UDP DNS responder on port 53.
- BleManager: NUS GATT service, sync_channel(8) command queue,
  dedicated notify thread (8KB), 3-level zombie defense.
- HttpServer: captive portal (5 redirects + WiFi form), REST API
  (ping, status, command, valve, SSE, logs), WebUI (html, css, js).
  SSE: blocking handler with mpsc channel + httpd_resp_send_chunk.
- DNS builder as pure domain function with 4 unit tests.
- Transport SM: USB > BLE > Advertising with LED indication.
- Box::leak for 'static refs to satisfy fn_handler 'static bound.

Hardware-confirmed: SSE stream (curl -N), 60s no Guru Meditation,
HTTP server (17 routes).
Pending: AP mode (NVS erase), BLE (NimBLE patch + uncomment init).

Report: docs/plans/completed/26_06_30_phase4_network_report.md

+++ src/config.rs
@@ -9,10 +9,10 @@ pub const AP_NETMASK_BITS: u8 = 24;
 pub const AP_IP_ADDRESS: &str = "192.168.4.1";
 
 // ── WiFi Station ──────────────────────────────────────────────
-pub const STA_CONNECT_TIMEOUT_MS: u32 = 15_000;
-pub const STA_RECONNECT_INTERVAL_MS: u32 = 30_000;
-pub const STA_POLL_MS: u32 = 500;
-pub const STA_POST_CONNECT_DELAY_MS: u32 = 500;
+pub const STA_CONNECT_TIMEOUT_MS: u64 = 15_000;
+pub const STA_RECONNECT_INTERVAL_MS: u64 = 30_000;
+pub const STA_POLL_MS: u64 = 500;
+pub const STA_POST_CONNECT_DELAY_MS: u64 = 500;
 
 // ── GPIO Pin Assignments ──────────────────────────────────────
 pub const PIN_STEP: u8 = 25;
